telegram-bot/requirements.txt
python-telegram-bot
fastapi
uvicorn
pydantic
httpx
python-dotenv
pytest
pytest-asyncio
requests
pydantic-settings
google-cloud-pubsub
